Chance's comments
