+++ sematic/ui/packages/common/src/typeViz/views/HuggingFaceReference.tsx
@@ -1,7 +1,8 @@
 import { OpenInNew } from "@mui/icons-material";
 import { Button, Tooltip } from "@mui/material";
+import { useMemo } from "react";
 import { useTextSelection } from "@sematic/common/src/hooks/textSelectionHooks";
-import { ViewComponentProps} from "src/typeViz/common";
+import { ViewComponentProps, ValueComponentProps} from "src/typeViz/common";
 
 type Reference = {
     owner: string | null;
@@ -10,7 +11,7 @@ type Reference = {
     subset?: string | null | undefined;
 }
 
-export function HuggingFaceModelReferenceShortView(props: ViewComponentProps) {
+export function HuggingFaceModelReferenceShortView(props: ValueComponentProps) {
     const { valueSummary } = props;
     const { values } = valueSummary;
 
@@ -21,7 +22,7 @@ export function HuggingFaceModelReferenceShortView(props: ViewComponentProps) {
     />;
 }
 
-export function HuggingFaceDatasetReferenceShortView(props: ViewComponentProps) {
+export function HuggingFaceDatasetReferenceShortView(props: ValueComponentProps) {
     const { valueSummary } = props;
     const { values } = valueSummary;
 
@@ -63,44 +64,54 @@ function HuggingFaceButton(props: {
     const { reference, objectTypePrefix, short } = props;
     const { owner, repo, subset, commit_sha } = reference;
 
-    const ownerSegment = owner ? "/" + owner : "";
-    const repoPath = objectTypePrefix + ownerSegment + "/" + repo;
+    const displayedHref = useMemo(
+        () => {
+            const ownerSegment = owner ? "/" + owner : "";
+            const repoPath = objectTypePrefix + ownerSegment + "/" + repo;
 
-    // hugging face doesn't allow you to link to a particular
-    // commit and subset in the UI at the same time. If there
-    // is a specific commit referenced, let that take precedence.
-    let path = repoPath;
-    if(commit_sha) {
-        path = repoPath + "/tree/" + commit_sha;
-    } else if(subset) {
-        path = repoPath + "/viewer/" + subset;
-    }
+            // hugging face doesn't allow you to link to a particular
+            // commit and subset in the UI at the same time. If there
+            // is a specific commit referenced, let that take precedence.
+            let path = repoPath;
+            if(commit_sha) {
+                path = repoPath + "/tree/" + commit_sha;
+            } else if(subset) {
+                path = repoPath + "/viewer/" + subset;
+            }
 
-    const href = new URL("https://huggingface.co" + path);
+            const href = new URL("https://huggingface.co" + path);
+            return href;
+        }, [owner, repo, subset, commit_sha, objectTypePrefix]
+    );
 
-    let slug = (owner ? owner + "/" : "") + repo;
-    if(subset) {
-        slug += ":" + subset;
-    }
-    if(commit_sha) {
-        slug += "@" + commit_sha.substring(0, 7);
-    }
+    const displayedSlug = useMemo(
+        () => {
+            let slug = (owner ? owner + "/" : "") + repo;
+            if(subset) {
+                slug += ":" + subset;
+            }
+            if(commit_sha) {
+                slug += "@" + commit_sha.substring(0, 7);
+            }
   
+            if(short) {
+                slug = "";
+            }
+            return slug;
+        }, [owner, repo, subset, commit_sha, short]
+    );
     const elementRef = useTextSelection<HTMLDivElement>();
-    if(short) {
-        slug = "";
-    }
 
     return <Tooltip title="View on Hugging Face">
         <Button
-            href={href.href}
+            href={displayedHref.href}
             variant="outlined"
             target="blank"
             endIcon={<OpenInNew />}
             draggable={false}
             style={{ userSelect: "text" }}
         >
-            <div ref={elementRef} style={{ cursor: "text" }} >🤗 {slug}</div>
+            <div ref={elementRef} style={{ cursor: "text" }} >🤗 {displayedSlug}</div>
         </Button>
     </Tooltip>;
 }
