+++ sematic/ui/packages/common/src/typeViz/views/PromptResponse.tsx
