+++ sematic/ui/packages/common/src/typeViz/vizMapping.ts
@@ -12,12 +12,12 @@ import {
     HuggingFaceDatasetReferenceValueView,
     HuggingFaceModelReferenceShortView,
     HuggingFaceModelReferenceValueView
-} from "src/typeViz/views/hugging_face_reference";
-import { HuggingFaceStoredModelShortView, HuggingFaceStoredModelFullView } from "src/typeViz/views/hugging_face_stored_model";
+} from "src/typeViz/views/HuggingFaceReference";
+import { HuggingFaceStoredModelShortView, HuggingFaceStoredModelFullView } from "src/typeViz/views/HuggingFaceStoredModel";
 import ImageValueView, { ImageExpandedView } from "src/typeViz/views/image";
 import IntValueView from "src/typeViz/views/int";
 import LinkValueView from "src/typeViz/views/link";
-import PromptResponseCollapsedView, { PromptResponseExpandedView } from "src/typeViz/views/prompt_response";
+import PromptResponseCollapsedView, { PromptResponseExpandedView } from "src/typeViz/views/PromptResponse";
 import ListValueView, { ListElementsView } from "src/typeViz/views/list";
 import NoneValueView from "src/typeViz/views/none";
 import PlotlyFigureValueView, { PlotlyFigureExpandedView } from "src/typeViz/views/plotly";
