+++ sematic/ui/packages/main/src/types/Types.tsx
@@ -12,15 +12,15 @@ import DatetimeValueView from "src/types/datatime";
 import DictValueView from "src/types/dict";
 import EnumValueView from "src/types/enum";
 import FloatValueView from "src/types/float";
-import { HuggingFaceDatasetReferenceValueView, HuggingFaceModelReferenceValueView } from "src/types/hugging_face_reference";
+import { HuggingFaceDatasetReferenceValueView, HuggingFaceModelReferenceValueView } from "src/types/HuggingFaceReference";
 import ImageValueView from "src/types/image";
 import IntValueView from "src/types/int";
 import LinkValueView from "src/types/link";
 import ListValueView from "src/types/list";
 import NoneValueView from "src/types/none";
 import PlotlyFigureValueView from "src/types/plotly";
-import PromptResponseValueView from "src/types/prompt_response";
-import HuggingFaceStoredModelView from "src/types/hugging_face_stored_model";
+import PromptResponseValueView from "src/types/PromptResponse";
+import HuggingFaceStoredModelView from "src/types/HuggingFaceStoredModel";
 import StrValueView from "src/types/str";
 import TorchDataLoaderValueView from "src/types/torchDataLoader";
 import TupleValueView from "src/types/tuple";
